requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asyncpg==0.30.0
attrs==24.3.0
audioop-lts==0.2.1
certifi==2024.8.30
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pathlib==1.0.1
pillow==11.0.0
propcache==0.2.1
pytesseract==0.3.13
python-dotenv==1.0.1
yarl==1.18.3
